9336 Remove public and container_state from repository
