--- repository/container_state.go
@@ -1,6 +1,0 @@
-package repository
-
-type ContainerState struct {
-	State  string `json:"state"`
-	Status string `json:"status"`
-}
