--- repository/public.go
@@ -1,9 +1,0 @@
-package repository
-
-type Public struct {
-	Enabled  bool     `json:"enabled"`
-	Hosts    []string `json:"hosts" binding:"required"`
-	Port     string   `json:"port" binding:"required,numeric"`
-	SSL      bool     `json:"ssl"`
-	Compress bool     `json:"compress"`
-}
